发布资源清单下载、高级搜索、合规趋势统计等API.

+++ config/include/alibabacloud/config/model/ListAggregateResourceRelationsResult.h
@@ -37,7 +37,7 @@ namespace AlibabaCloud
 					struct ResourceRelationListItem
 					{
 						std::string targetResourceType;
-						int accountId;
+						long accountId;
 						std::string targetResourceId;
 						std::string relationType;
 						std::string sourceResourceRegionId;
